Don't compare jiffy and java result if they can't be compared
If jiffy has returned something other than an exception (e.g. call to jiffy
succeeded) and the corresponding call to Java resulted in an exception,
don't call `same?`, as the implementations of `same?` excepts the arguments
to be of the same "kind", i.e. either both exception or both something other
than exception.